refactor: Facade 패턴을 이용한 S3 I/O 및 DB 트랜잭션 분리
- GroupService를 Facade, CommandService, QueryService로 3단 분리하여 역할 분담
- S3 업로드를 트랜잭션 외부로 이동해 DB 커넥션 점유 시간 단축 (커넥션 풀 고갈 방지)
- DB 예외 발생 시 S3 이미지 삭제 보상 로직 추가로 스토리지-DB 간 데이터 정합성 확보
